// src/components/ContactForm/ContactForm.types.ts
export enum InputName {
  NAME = "name",
  EMAIL = "email",
  MESSAGE = "message",
}
export enum InputType {
  NAME = "text",
  EMAIL = "text",
  MESSAGE = "textarea",
}
export enum InputLabels {
  NAME = "imię i nazwisko",
  EMAIL = "twój email",
  MESSAGE = "wiadomość",
}
export enum ErrorMsg {
  NAME_REQ = "podaj swoje imię",
  EMAIL_REQ = "podaj swój email",
  EMAIL_WRG = "podaj poprawny email",
  MESS_SHRT = "minimum 10 znaków",
  MESS_REQ = "napisz coś do nas!",
}